+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 - Give access to configlet to Site Administrator [sgeulette].
+- Don't store None in catalog [sgeulette]
 
 
 0.2 (2015-06-02)